src/presentation/controllers/signup/signup.ts
import { Controller, HttpRequest, HttpResponse, Validation } from './signup-protocols'
import { badRequest, okResponse, serverError } from '../../helpers'
import { AddAccount } from '../../../domain/usecases/add-account'

export class SignUpController implements Controller {
  private readonly addAccount: AddAccount
  private readonly validation: Validation

  constructor (addAccount: AddAccount, validation: Validation) {
    this.addAccount = addAccount
    this.validation = validation
  }

  async handle (httpRequest: HttpRequest): Promise<HttpResponse> {
    try {
      const error = await this.validation.validate(httpRequest.body)
      if (error) return badRequest(error)

      const { name, email, password } = httpRequest.body
      const account = await this.addAccount.add({
        name,
        email,
        password
      })

      return okResponse(account)
    } catch (error) {
      return serverError(error)
    }
  }
}
